SMDR/requirements.txt
appdirs==1.4.3
asgiref==3.3.1
astroid==2.4.2
autopep8==1.6.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.0
distro==1.4.0
Django==3.0
django-cleanup==5.1.0
django-cors-headers==3.5.0
django-countries==6.1.3
django-filter==2.4.0
django-jazzmin==2.4.1
django-mdeditor==0.1.20
django-summernote==0.8.11.6
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
honeybadger==0.8.0
html5lib==1.0.1
idna==2.8
inflection==0.5.1
ipaddr==2.2.0
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
progress==1.5
psutil==5.9.1
pycodestyle==2.8.0
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.6
python-decouple==3.3
pytoml==0.1.21
pytz==2020.4
requests==2.24.0
retrying==1.3.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.14.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
wrapt==1.12.1
